backend-server/pyproject.toml
[tool.poetry]
name = "backend-server"
version = "0.0.1"
description = "Nevermind chat backend server"
authors = ["Jovian Dsouza <dsouzajovian123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.335"
litellm = "^0.14.1"
openai = "^1.2.4"
python-dotenv = "^1.0.0"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"